requirements.txt
Django<1.9,>=1.8
django-celery
lxml
requests
